Разделение на модули (правки)
