+++ js/main.js
@@ -1,4 +1,3 @@
-import './data.js';
-import {generatePhoto} from './data.js';
+import {generatePhoto} from './generate-photo.js';
 
 generatePhoto(2);
